+++ jsapp/js/account/usage/usageTopTabs.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react'
 
 import { useNavigate } from 'react-router-dom'
 import { ACCOUNT_ROUTES } from '#/account/routes.constants'
-import TabsOld from '#/components/common/TabsOld'
+import Tabs from '#/components/common/tabs'
 import Usage from './usage.component'
 import ProjectBreakdown from './usageProjectBreakdown'
 
@@ -21,7 +21,7 @@ const usageTopTabs: React.FC<UsageTopTabsProps> = ({ activeRoute }) => {
 
   return (
     <div>
-      <TabsOld
+      <Tabs
         tabs={[
           { label: t('Account Total'), route: ACCOUNT_ROUTES.USAGE },
           {
