+++ jsapp/js/theme/kobo/index.ts
@@ -20,7 +20,6 @@ import { PillThemeKobo } from './Pill'
 import { RadioThemeKobo } from './Radio'
 import { SelectThemeKobo } from './Select'
 import { TableThemeKobo } from './Table'
-import { TabsThemeKobo } from './Tabs'
 import { TagsInputThemeKobo } from './TagsInput'
 import { TextareaThemeKobo } from './Textarea'
 import { ThemeIconThemeKobo } from './ThemeIcon'
@@ -167,7 +166,6 @@ export const themeKobo = createTheme({
     Divider: DividerThemeKobo,
     Dropzone: DropzoneThemeKobo,
     TagsInput: TagsInputThemeKobo,
-    Tabs: TabsThemeKobo,
     ThemeIcon: ThemeIconThemeKobo,
     NumberInput: NumberInputThemeKobo,
     Textarea: TextareaThemeKobo,
